+++ src/anemoi/registry/commands/weights.py
@@ -8,10 +8,6 @@
 # nor does it submit to any jurisdiction.
 
 
-"""Command place holder. Delete when we have real commands.
-
-"""
-
 import logging
 
 from ..entry.weights import WeightCatalogueEntry
@@ -39,11 +35,15 @@ class Weights(BaseCommand):
             action="store_true",
         )
         # command_parser.add_argument("--delete", help=f"Delete the {self.kind} from the catalogue and from any other location", action="store_true")
+        self.add_set_get_remove_metadata_arguments(command_parser)
 
         command_parser.add_argument("--add-location", help="Platform to add location to the weights.")
         command_parser.add_argument("--location-path", help="Path of the new location using {{uuid}}.", metavar="PATH")
         command_parser.add_argument("--overwrite", help="Overwrite any existing weights.", action="store_true")
         command_parser.add_argument("--url", help="Print the URL of the dataset.", action="store_true")
+        command_parser.add_argument(
+            "--view", help=f"Open the URL of the {self.kind} in a browser.", action="store_true"
+        )
 
     def _run(self, entry, args):
         if args.unregister:
@@ -51,8 +51,14 @@ class Weights(BaseCommand):
             return
         self.process_task(entry, args, "register", overwrite=args.overwrite)
         self.process_task(entry, args, "add_location", path=args.location_path)
+        self.set_get_remove_metadata(entry, args)
+
         if args.url:
             print(entry.url)
+        if args.view:
+            import webbrowser
+
+            webbrowser.open(entry.url)
 
 
 command = Weights