+++ src/anemoi/registry/entry/dataset.py
@@ -9,8 +9,10 @@
 
 
 import datetime
+import glob
 import logging
 import os
+import shutil
 
 import yaml
 from anemoi.datasets import open_dataset
@@ -25,6 +27,40 @@ from . import CatalogueEntry
 LOG = logging.getLogger(__name__)
 
 COLLECTION = "datasets"
+
+
+def delete_on_s3(path):
+    from anemoi.utils.remote.s3 import delete
+
+    return delete(path + "/")
+
+
+def delete_on_local(path):
+    if not os.path.exists(path):
+        LOG.warning(f"Path {path} does not exist, nothing to delete.")
+        return
+    if not os.path.isdir(path):
+        raise ValueError(f"Path {path} is not a directory, cannot delete.")
+
+    # First, move the directory to a temporary name <dataset>.deleting.i
+    # This is to make sure the dataset is not accessed while we are deleting it.
+    tmp_path = path + ".deleting"
+    i = 0
+    while os.path.exists(tmp_path):
+        i += 1
+        tmp_path = path + ".deleting." + str(i)
+    try:
+        os.rename(path, tmp_path)
+    except OSError as e:
+        LOG.error(f"Failed to rename {path} to {tmp_path}: {e}")
+        raise
+
+    # Now do the actual deletion, of any <dataset>.deleting* directories
+    for to_delete in glob.glob(path + ".deleting*"):
+        try:
+            shutil.rmtree(to_delete)
+        except OSError as e:
+            LOG.error(f"Failed to delete {to_delete}: {e}")
 
 
 class DatasetCatalogueEntryList(RestItemList):
@@ -64,7 +100,7 @@ class DatasetCatalogueEntry(CatalogueEntry):
         entry.add_location(PLATFORM, target)
 
     def set_status(self, status):
-        self.patch([{"op": "add", "path": "/status", "value": status}])
+        self.patch([{"op": "add", "path": "/status", "value": status}], robust=True)
 
     def build_location_path(self, platform, uri_pattern=None):
         if uri_pattern is None:
@@ -81,26 +117,31 @@ class DatasetCatalogueEntry(CatalogueEntry):
             path = os.path.normpath(path)
 
         LOG.debug(f"Adding location to {platform}: {path}")
-        self.patch([{"op": "add", "path": f"/locations/{platform}", "value": {"path": path}}])
+        self.patch([{"op": "add", "path": f"/locations/{platform}", "value": {"path": path}}], robust=True)
         return path
 
     def remove_location(self, platform):
-        self.patch([{"op": "remove", "path": f"/locations/{platform}"}])
+        self.patch([{"op": "remove", "path": f"/locations/{platform}"}], robust=True)
+        LOG.warning(f"Removed location from catalogue from '{platform}'")
 
     def delete_location(self, platform):
+        # the variable name 'location' and 'platform' are sometimes used interchangeably in the codebase
+        # this should be clarified in the future
+
         if not config().get("allow_delete"):
             raise ValueError("Delete not allowed by configuration")
 
         path = self.record.get("locations", {}).get(platform, {}).get("path")
         if path is None:
-            LOG.warning(f"Nothing to delete for {self.key} on platform {platform}")
+            LOG.warning(f"ERROR: Nothing to delete for {self.key} on platform {platform}")
             return
         if path.startswith("s3://"):
-            from anemoi.utils.remote.s3 import delete
-
-            return delete(path + "/")
+            delete_on_s3(path)
         else:
-            LOG.warning(f"Location is not an s3 path: {path}. Delete not implemented.")
+            LOG.warning(f"Deleting {path} from '{platform}'")
+            delete_on_local(path)
+
+        LOG.warning(f"Deleted {path} from '{platform}'")
 
         self.remove_location(platform)
 
@@ -168,11 +209,11 @@ class DatasetCatalogueEntry(CatalogueEntry):
 
     def set_recipe(self, file):
         recipe = self._file_or_dict(file)
-        self.patch([{"op": "add", "path": "/metadata/recipe", "value": sanitise(recipe)}])
+        self.patch([{"op": "add", "path": "/metadata/recipe", "value": sanitise(recipe)}], robust=True)
 
     def set_variables_metadata(self, file):
         variables_metadata = self._file_or_dict(file)
-        self.patch([{"op": "add", "path": "/metadata/variables_metadata", "value": variables_metadata}])
+        self.patch([{"op": "add", "path": "/metadata/variables_metadata", "value": variables_metadata}], robust=True)
 
     def load_from_path(self, path):
         import zarr