+++ src/anemoi/registry/rest.py
@@ -15,9 +15,8 @@ import socket
 from getpass import getuser
 
 import requests
+from anemoi.utils.remote import robust as make_robust
 from requests.exceptions import HTTPError
-
-from anemoi.registry import config
 
 from ._version import __version__
 
@@ -66,15 +65,28 @@ def trace_info():
 class Rest:
     """REST API client."""
 
-    def __init__(self):
+    def __init__(self, token=None):
+        self.token = token or self.config.api_token
+
         self.session = requests.Session()
         self.session.headers.update({"Authorization": f"Bearer {self.token}"})
         for k, v in trace_info().items():
             self.session.headers.update({f"x-anemoi-registry-{k}": str(v)})
 
     @property
-    def token(self):
-        return config().api_token
+    def config(self):
+        from anemoi.registry import config
+
+        return config()
+
+    @property
+    def api_url(self):
+        return self.config.api_url
+
+    def get_url(self, url):
+        r = make_robust(self.session.get)(url)
+        self.raise_for_status(r)
+        return r.json()
 
     def get(self, path, params=None, errors={}):
         self.log_debug("GET", path, params)
@@ -83,7 +95,7 @@ class Rest:
         if params is not None:
             kwargs["params"] = params
 
-        r = self.session.get(f"{config().api_url}/{path}", **kwargs)
+        r = make_robust(self.session.get)(f"{self.api_url}/{path}", **kwargs)
         self.raise_for_status(r, errors=errors)
         return r.json()
 
@@ -99,30 +111,37 @@ class Rest:
         self.log_debug("PUT", path, data)
         if not data:
             raise ValueError(f"PUT data must be provided for {path}")
-        r = self.session.put(f"{config().api_url}/{path}", json=tidy(data))
+        r = make_robust(self.session.put)(f"{self.api_url}/{path}", json=tidy(data))
         self.raise_for_status(r, errors=errors)
         return r.json()
 
-    def patch(self, path, data, errors={}):
+    def patch(self, path, data, errors={}, robust=False):
+        # patch (and post) are not idempotent, so we need to be careful with retries
+        # default to non-robust
+        robust_ = {True: make_robust, False: lambda x: x}[robust]
+
         self.log_debug("PATCH", path, data)
         if not data:
             raise ValueError(f"PATCH data must be provided for {path}")
-        r = self.session.patch(f"{config().api_url}/{path}", json=tidy(data))
+        r = robust_(self.session.patch)(f"{self.api_url}/{path}", json=tidy(data))
         self.raise_for_status(r, errors=errors)
         return r.json()
 
-    def post(self, path, data, errors={}):
-        r = self.session.post(f"{config().api_url}/{path}", json=tidy(data))
+    def post(self, path, data, errors={}, robust=False):
+        # patch (and post) are not idempotent, so we need to be careful with retries
+        robust_ = {True: make_robust, False: lambda x: x}[robust]
+
+        r = robust_(self.session.post)(f"{self.api_url}/{path}", json=tidy(data))
         self.raise_for_status(r, errors=errors)
         return r.json()
 
     def delete(self, path, errors={}):
-        if not config().get("allow_delete"):
+        if not self.config.get("allow_delete"):
             raise ValueError("Unregister not allowed")
         return self.unprotected_delete(path, errors=errors)
 
     def unprotected_delete(self, path, errors={}):
-        r = self.session.delete(f"{config().api_url}/{path}", params=dict(force=True))
+        r = make_robust(self.session.delete)(f"{self.api_url}/{path}", params=dict(force=True))
         self.raise_for_status(r, errors=errors)
         return r.json()
 
@@ -164,28 +183,28 @@ class RestItem:
         self.rest = Rest()
         self.path = f"{collection}/{key}"
 
-    def exists(self):
+    def exists(self, *args, **kwargs):
         try:
-            self.get()
+            self.get(*args, **kwargs)
             return True
         except HTTPError as e:
             if e.response.status_code == 404:
                 return False
 
-    def get(self, params=None, errors={}):
-        return self.rest.get(self.path, params=params, errors=errors)
+    def get(self, *args, **kwargs):
+        return self.rest.get(self.path, *args, **kwargs)
 
-    def patch(self, data):
-        return self.rest.patch(self.path, data)
+    def patch(self, data, *args, **kwargs):
+        return self.rest.patch(self.path, data, *args, **kwargs)
 
-    def put(self, data):
-        return self.rest.put(self.path, data)
+    def put(self, data, *args, **kwargs):
+        return self.rest.put(self.path, data, *args, **kwargs)
 
-    def delete(self):
-        return self.rest.delete(self.path)
+    def delete(self, *args, **kwargs):
+        return self.rest.delete(self.path, *args, **kwargs)
 
-    def unprotected_delete(self):
-        return self.rest.unprotected_delete(self.path)
+    def unprotected_delete(self, *args, **kwargs):
+        return self.rest.unprotected_delete(self.path, *args, **kwargs)
 
     def __repr__(self):
         return f"{self.__class__.__name__}({self.collection}, {self.key})"
@@ -199,11 +218,11 @@ class RestItemList:
         self.rest = Rest()
         self.path = collection
 
-    def get(self, params=None, errors={}):
-        return self.rest.get(self.path, params=params, errors=errors)
+    def get(self, *args, **kwargs):
+        return self.rest.get(self.path, *args, **kwargs)
 
-    def post(self, data):
-        return self.rest.post(self.path, data, errors={409: AlreadyExists})
+    def post(self, data, **kwargs):
+        return self.rest.post(self.path, data, errors={409: AlreadyExists}, **kwargs)
 
     def __repr__(self):
         return f"{self.__class__.__name__}({self.collection})"