+++ CHANGELOG.md
@@ -8,6 +8,58 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please add your functional changes to the appropriate section in the PR.
 Keep it human-readable, your future self will thank you!
 
+## [0.2.2](https://github.com/ecmwf/anemoi-registry/compare/0.2.1...0.2.2) (2025-06-30)
+
+
+### Features
+
+* Add trainings subcommand ([#70](https://github.com/ecmwf/anemoi-registry/issues/70)) ([1f92d0b](https://github.com/ecmwf/anemoi-registry/commit/1f92d0b9965588823aea940ff055e26001efd637))
+* Implement worker to delete datasets on s3 ([#85](https://github.com/ecmwf/anemoi-registry/issues/85)) ([8de1f71](https://github.com/ecmwf/anemoi-registry/commit/8de1f710bae2abed017e1e9b9ed51affb4c27587))
+* Simplify getting settings from catalogue ([#81](https://github.com/ecmwf/anemoi-registry/issues/81)) ([cada143](https://github.com/ecmwf/anemoi-registry/commit/cada143bb61e9c806c8e9100074adda860a0d994))
+* Transfering datasets improvement ([#93](https://github.com/ecmwf/anemoi-registry/issues/93)) ([c33330e](https://github.com/ecmwf/anemoi-registry/commit/c33330e61536b0d6146cb727c6418a8d2318a8f9))
+* Use settings from server ([#77](https://github.com/ecmwf/anemoi-registry/issues/77)) ([1633674](https://github.com/ecmwf/anemoi-registry/commit/1633674ce4004098618c32ffcbdc7c916cccbce2))
+
+
+### Bug Fixes
+
+* Datasets deletion improvement ([#94](https://github.com/ecmwf/anemoi-registry/issues/94)) ([50700b6](https://github.com/ecmwf/anemoi-registry/commit/50700b6f82fc8f1761bad7edebb98d6924afe8f9))
+* Fix deleting datasets ([#95](https://github.com/ecmwf/anemoi-registry/issues/95)) ([64e70c8](https://github.com/ecmwf/anemoi-registry/commit/64e70c800dd2bcff78b83e01ff536040493aca78))
+
+
+### Documentation
+
+* Minor updates ([#83](https://github.com/ecmwf/anemoi-registry/issues/83)) ([ffeabc5](https://github.com/ecmwf/anemoi-registry/commit/ffeabc5f8cae118daf6ff8057dfca5b3e2846159))
+
+## [0.2.1](https://github.com/ecmwf/anemoi-registry/compare/0.2.0...0.2.1) (2025-05-20)
+
+
+### Features
+
+* use latest version of anemoi-datasets ([#76](https://github.com/ecmwf/anemoi-registry/issues/76)) ([aa4ab87](https://github.com/ecmwf/anemoi-registry/commit/aa4ab87486381ba4d8fcc56d15c94d87adcb729d))
+
+
+### Bug Fixes
+
+* remove pinning on botocore and update anemoi-utils ([7665a23](https://github.com/ecmwf/anemoi-registry/commit/7665a23112e61f91aed0e657aab37146967dc40a))
+
+## [0.2.0](https://github.com/ecmwf/anemoi-registry/compare/0.1.8...0.2.0) (2025-04-10)
+
+
+### Features
+
+* add robust connection to server ([#66](https://github.com/ecmwf/anemoi-registry/issues/66)) ([c0f735c](https://github.com/ecmwf/anemoi-registry/commit/c0f735ccf132e924c476539e3ae808575c60c324))
+* adding set-get-remove metadata ([#63](https://github.com/ecmwf/anemoi-registry/issues/63)) ([15633fe](https://github.com/ecmwf/anemoi-registry/commit/15633feb2bf9664b64026729aa763c666991c6e3))
+
+
+### Documentation
+
+* new logo ([#69](https://github.com/ecmwf/anemoi-registry/issues/69)) ([4e40632](https://github.com/ecmwf/anemoi-registry/commit/4e406322e3a74bb9263fb791523b41fce10508ac))
+
+
+### Miscellaneous Chores
+
+* Manual version bump ([c2f0b8f](https://github.com/ecmwf/anemoi-registry/commit/c2f0b8f26deb089385df8ba5b35cc5b69ea06ffa))
+
 ## [0.1.8](https://github.com/ecmwf/anemoi-registry/compare/0.1.7...0.1.8) (2025-03-04)